feat: add migration tips
